+++ src/apis/auth/authApi.ts
@@ -1,4 +1,4 @@
-import { ApiResponse } from "@/types/common/api.type";
+import { ApiResponse } from "@/types/common/apiResponse.type";
 import {
   PostLoginRequest,
   PostLoginResponse,
