+++ src/apis/team/teamApi.ts
@@ -1,4 +1,4 @@
-import { ApiResponse } from "@/types/apiResponse.type";
+import { ApiResponse } from "@/types/common/apiResponse.type";
 import {
   GetDivisionResponse,
   GetGroupResponse,
